chore: refactor out the route into its own folder
